Simple UI changes and bugs related to navigation fixed
Removed:
Implication of the basic navigation.
Use of local storage folder for further research.
Doing:
Implementing the gyrosensors and accelerometer sensors and extracting
the data. Going to be uploaded in next commit.
TO-DO:
Use the local storage cache for data state preservation and export the
file instead of text file to CSV file.